requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
joblib==1.3.2
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
scipy==1.11.4
threadpoolctl==3.2.0
tk==0.1.0
ttkthemes==3.2.2
